Program to pull information from companies' ERP system and autofill information on a manufacturing drawing created in SolidWorks
When I would create technical drawings to be used in our manufacturing facility to build a product, one of the processes involved manually looking up critical information in our ERP database to place on the drawing. This was tedious and very time consuming, and at times could cause misinformation due to user error. I developed this program in Visual Basic to eliminate errors and reduce all of the manual work that was required in the past to a single macro. Please see the read me file in the main branch for a detailed explanation on the process of the program.