Create Pull information from a database and use a batch script to convert 8000 outdated files into pdf format
This program was used, in conjunction with another program, to batch convert over 8000 drawings into pdf format. The existing program that was currently being used would take an input from a user, and generate a custom drawing that currently existed as a pdf, which was in a directory on the server. It also created a custom report, based on information that was stored in a text file, using variables pulled from the ERP database. 

The downside to this method was that the program had to be run each time a user wanted to view a custom drawing/report. The program was run on Windows XP, and the company that created it had gone out of business. The program would also constantly experience issues, and crash more often then not.

Running this program allowed me to loop through all record variables of a SQL query, that searched for items in the database that had an distinct variable. This variable represented that the record had an associated custom drawing that currently existed on the server. If this variable was null, it would skip over that record and move to the next.

This allowed me to generate a text file, that the outdated program normally generated individually, for all items with an associated drawing. The outdated program supported batch scripting, so I also wrote a batch file that would simulate the steps the program took to generate the custom drawing when entered manually on an individual basis.

Using these two programs, I was able to convert over 8000 of these drawings to pdf format, and set up a version control system that allowed for revisions to the files. This saved the company thousands of dollars compared to a quote we received from an outside company to solve the same problem.