Create Batch file to create a new directory and sub-directories for new customer orders
When receiving a new order from a customer, we would create a folder on our server, that had a specific job number related to that customer. Many times this was done manually, which led to an inconsistent folder structure, and misplaced files.

This allowed a user to create a new directory based on inputs for the folder name, and the fiscal year. It also copied templates that I had created to perform engineering calculations and track progress of the project. This led to a consistent folder structure that was the same across all files, and made finding information much quicker and easier.