Update Program to pull information from companies' ERP system and autofill information on a manufacturing drawing
This was a program I had written which utilized both Visual Basic and SQL to automatically generate a technical drawing from a 3D design. Due to information existing in multiple locations across the company, engineers were required to manually search for and enter values into fields such as a bill of materials. This led to inconsistencies and errors across documents, and caused production slowdowns as data had to be verified and corrected prior to releasing documents to the production floor.

The goal was to create a consistent system of technical drawings by automatically entering required values based only on the name of the drawing. As all of the information that was required currently existed on the SQL server, a query could be utilized to pass through values to variables defined in the macro, and linked to empty fields in the document.

In order for this macro to run correctly, the data had to either exist in the database or be entered prior to creating the technical drawing. As this was the current company procedure, it was not much of a concern, however there was an effort to clean the database prior to this project. Error checking in the form of message boxes was also implemented in the program at specific steps, in order to visually see if there were any issues during execution.

The function of the program is as follows:

1. Declare variables for SolidWorks API, database connection, SQL queries, part, and part component values
2. Obtain the name of the drawing, which corresponds to the part to be looked up in the ERP system
3. Declare variables for the Part, and the Sub-Parts that it is made up of (Subordinate Parts)
4. Build the SQL queries, using the part name as the Primary ID
5. Open and verify database connection
6. Run SQL queries and pass them to a record set variable to be evaluated
7. Verify the record sets are not empty
8. Check to see if there are Sub-Parts that exist that make up the main Part
9. If they exist, run a sub-query and loop through the record sets
10. Find the correct matching record sets based on information from the previously run queries (as multiple records can exist and may not be correct)
11. Store the values in a string array
12. Empty the custom property fields on the technical drawing (in case information exists)
13. Pass through the information to custom property fields on the technical drawing
14. Empty the record set variables to avoid data contamination
15. Close the database connections