Macro Functions - Custom Excel Purchase Order Form
This was a group of macro functions that I had written in order to speed up data entry when utilizing company purchase order forms. The initial problem was that the procedure for filling out one of these forms involved multiple manual operations, including opening a master sheet of capital expenditures to obtain a purchase order number.

Since this was a manual operation, there were times when users forgot to input the correct purchase order number, or used a previously existing number, leading to confusion when reviewing CapEx for end-of-month expenses.

A user would also have to manually generate a pdf file when the form was completed, and then save a copy of the excel file in a directory based on the current year. The filename would also have to reflect the purchase order number, which was inside the excel file.

My solution was to place these macros inside a master template file in the form of large buttons, which allowed the user to easily access each function. When a user had completed the form and all procedures were followed, they could then close the file. Upon closing the file, the template would also reset to a default state, however, it would prompt the user to make sure that they had saved a copy with their information in the correct location.