Query ERP Database and Update Legacy CAD Files
This project was initiated due to the company utilizing a legacy CAD software program that was not compatible after a complete server upgrade. The legacy program functioned by pulling in data from the companies' ERP system into a text file, and auto-filling text fields on a stationary title block template. It also inserted a DXF image into the drawing that already existed in another location on the server.

As this was an older program, the company that had designed it was no longer in business. I had to reverse-engineer the operations of the program and utilize a combination of VB and SQL to replicate the function. The goal was to run this as a batch file in a virtual environment where the program still worked, and convert all legacy drawings into PDF files that were revision-controlled in the company's Product Data Management system.

The first step was to determine the framework of the program. Some items that had to be considered were as follows:
1. Many parts were made from "Sub-Parts" called an "Engineering Master" that existed in various table locations in the ERP database
2. There could be multiple "Engineering Masters" for the same part
3. The "Engineering Masters" could have multiple "Sub Parts" as well
4. The data fields in the recordsets could be missing

The solution was to generate SQL queries that used the top-level part name, or "Drawing_ID" as the clause. It would then loop through the results and write them to a notepad file template defined inside the program. 

After obtaining this information for all parts in the database (around 10,000 files) there was another batch script created that would call the legacy program, and feed it the information gathered in these text files to generate a drawing, then output the result as a pdf. This was used in a loop to open the program, generate the drawing, save it as a pdf, and close the legacy program, for each of the 10,000 files.