Batch File for Quotes and Orders - Standard File Structure
This was a batch file that I had created that standardized the creation of project folders for new quotes and orders. Due to many individuals entering quotes and orders into the system, there was a lack of consistency in naming convention and file structure.

This allowed for uniformity across all projects and made searching through the folder structure easier to navigate.